requirements.txt
djangorestframework==3.11.1
drf-yasg==1.17.1